metaresult: ensure consistent orphan meta cleanup
Delete a Meta object if it becomes orphaned after its result link is
removed.

+++ bublik/interfaces/signals.py
@@ -13,6 +13,7 @@ from bublik.data.models import (
     ConfigTypes,
     GlobalConfigs,
     Meta,
+    MetaResult,
     MetaTest,
     TestIterationResult,
 )
@@ -59,8 +60,12 @@ def categorize_metas_on_config_change(sender, instance, **kwargs):
 
 
 @receiver(post_delete, sender=MetaTest)
+@receiver(post_delete, sender=MetaResult)
 def delete_orphan_meta(sender, instance, **kwargs):
-    if not MetaTest.objects.filter(meta_id=instance.meta_id).exists():
+    if (
+        not MetaTest.objects.filter(meta_id=instance.meta_id).exists()
+        and not MetaResult.objects.filter(meta_id=instance.meta_id).exists()
+    ):
         Meta.objects.filter(id=instance.meta_id).delete()
